dev(#23): use new katex package for markdown-it
The old KaTeX package for markdown-it, `markdown-it-katex`, is fairly
outdated and no longer maintained. Therefore, some modern KaTeX features
are not available. There exist more recent forks such as
`@iktakahiro/markdown-it-katex`. These use more recent version of KaTeX
but are also not actively maintained.

Instead, use `markdown-it-texmath`, which uses a very recent version of
KaTeX and seems to be reasonably well maintained and tested.