feat(#177): enable understanding of front matter
**Why:**

Some users add metadata to their Markdown files in the form of
YAML-style front matter and it currently gets treated as regular
content, which looks a bit odd when rendered on the page.

Instead, we would like to be able to identify front matter and not
render it to the page. Ideally we would like to make the metadata
available and/or optionally render it to the page in some way, but this
would require some reworking of the current setup.

**What:**

* Install `markdown-it-front-matter`[^1], which enables `markdown-it` to
  parse YAML-style front matter.
* Enable `markdown-it-front-matter` via `MarkdownIt.use()` like we do
  with other plugins.
  * **Note:** This doesn’t actually do anything other than enable
    `markdown-it` to recognise front matter and not treat it as regular
    content.

[^1]: https://github.com/ParkSB/markdown-it-front-matter

---

https://github.com/jannis-baum/Vivify/issues/177