feat(#141): shorten alert CSS classname
Removed `classPrefix` config option, that's not really intended to be a
user's responsibility plus if it was user-configured, that would break
our default styles. Just more trouble than worth.

There's some concern whether class ".alert" is too generic and would
clash with something else, but probably not in this project.

Note: GitHub itself uses the class "markdown-alert", but we thought the
      "markdown-" prefix is quite ridiculous.