fix(#141): fix fallback alert styles
There was a mistake where if an icon is defined for a custom marker, it
no longer has the .fallback-alert class, so it doesn't even get the
basic structure of the alert.

Styles for the generic .alert need to be explicitly set.

To do this and *still* be able to set 'fallback' completely separately
from 'note', I added a new color variable:

    --alert-fallback

which also makes the fallback color easier to customize.