feat(#141): handle all sources of icons
Icons can be set one of the following ways:

1. octicon
2. filepath to an svg file
    - full path
    - full path with tilde
    - relative path (relative to Vivify config dir)
3. inline svg
    - unfortunately quite annoying in json format, but works when
      escaping quotes

Problem with custom icons, such as from lucide.dev:
    - octicons work best with `fill="currentColor"`,
    - but lucide icons unwantedly fill the entire icon with that style
        - these would prefer `fill="none" stroke="currentColor"`,
          but octicons work badly this way.

However, looking at the set of icons, I would greatly prefer if we
switched primary icon providers from octicon to lucide (or something
else, if there are even better alternatives). We'll consider that in a
followup issue.